+++ .stylelintrc.js
@@ -4,6 +4,13 @@ module.exports = {
   reportNeedlessDisables: true,
   reportInvalidScopeDisables: true,
   rules: {
-    'selector-pseudo-class-no-unknown': null
+    'selector-pseudo-class-no-unknown': null,
+    'max-nesting-depth': [
+      1,
+      {
+        ignore: ['pseudo-classes'],
+        ignoreAtRules: ['include']
+      }
+    ]
   }
 }