+++ services/web-app/lib/github.js
@@ -5,20 +5,26 @@
  * LICENSE file in the root directory of this source tree.
  */
 import yaml from 'js-yaml'
+import { isEmpty, isEqual } from 'lodash'
 import { getPlaiceholder } from 'plaiceholder'
-import slugify from 'slugify'
 
 import { IMAGES_CACHE_PATH } from '@/config/constants'
 import { libraryAllowList } from '@/data/libraries'
 import { getResponse, writeFile } from '@/lib/file-cache'
+import { getSlug } from '@/utils/slug'
 import { removeLeadingSlash } from '@/utils/string'
 
 /**
  * Validates the route's parameters and returns an object that also includes the library's slug as
- * well as path to the directory that contains the carbon-library.yml. Return an empty object if
+ * well as path to the directory that contains the carbon-library.yml. Returns an empty object if
  * not found. Does not validate ref, so people can set their own branch / tag / commit.
+ * @param {import('../typedefs').Params} params - Partially-complete parameters
+ * @returns {Promise<import('../typedefs').Params>} Complete parameters
  */
 const validateLibraryParams = async (params = {}) => {
+  /**
+   * @type {import('../typedefs').Params}
+   */
   let returnParams = {}
 
   for (const [slug, library] of Object.entries(libraryAllowList)) {
@@ -29,7 +35,7 @@ const validateLibraryParams = async (params = {}) => {
       params.library === slug
     ) {
       returnParams = {
-        slug,
+        library: slug,
         ...library
       }
 
@@ -45,29 +51,66 @@ const validateLibraryParams = async (params = {}) => {
 
   // get default branch if a branch isn't specified through params
 
-  if (!returnParams.ref) {
-    try {
-      const repo = await getResponse(returnParams.host, 'GET /repos/{owner}/{repo}', {
-        owner: returnParams.org,
-        repo: returnParams.repo
-      })
+  try {
+    const repo = await getResponse(returnParams.host, 'GET /repos/{owner}/{repo}', {
+      owner: returnParams.org,
+      repo: returnParams.repo
+    })
 
+    if (repo && !returnParams.ref) {
       returnParams.ref = repo.default_branch
-    } catch (err) {}
-  }
+    }
+  } catch (err) {}
 
   return returnParams
 }
 
 /**
+ * Takes an inheritance reference like `ibmdotcom-styles@latest/back-to-top` and returns parmas from
+ * the library allowlist.
+ * @param {string} inheritanceRef - Inheritance reference
+ * @returns {import('../typedefs').Params} Resource parameters
+ */
+const getParamsFromInheritedAsset = (inheritanceRef = '') => {
+  /**
+   * @type {import('../typedefs').Params}
+   */
+  let returnParams = {}
+
+  const [libraryId] = inheritanceRef.split('@')
+  const libraryRef = inheritanceRef.slice(
+    inheritanceRef.indexOf('@') + 1,
+    inheritanceRef.lastIndexOf('/')
+  )
+  const [assetId] = inheritanceRef.split('/').reverse()
+
+  returnParams = libraryAllowList[libraryId] || {}
+
+  if (libraryRef !== 'latest') {
+    returnParams.ref = libraryRef
+  }
+
+  return {
+    ...returnParams,
+    library: libraryId,
+    asset: assetId
+  }
+}
+
+/**
  * If the params map to a valid library in the allowlist, fetch the contents of the library's
  * metadata file. If the params are not valid, early return so the page redirects to 404.
+ * @param {import('../typedefs').Params} params
+ * @returns {import('../typedefs').Library}
  */
 export const getLibraryData = async (params = {}) => {
   const libraryParams = await validateLibraryParams(params)
 
-  if (!libraryParams || Object.keys(libraryParams).length === 0) return null
+  if (isEmpty(libraryParams)) return null
 
+  /**
+   * @type {import('../typedefs').GitHubContentResponse}
+   */
   let response = {}
 
   try {
@@ -81,19 +124,25 @@ export const getLibraryData = async (params = {}) => {
     return null
   }
 
+  /**
+   * @type {import('../typedefs').LibraryContent}
+   */
   const content = yaml.load(Buffer.from(response.content, response.encoding).toString())
 
-  const assets = await getLibraryAssets(params)
+  const assets = await getLibraryAssets(params, true)
+
+  const packageJsonContent = await getPackageJsonContent(params, content.packageJsonPath)
 
   const filteredAssets = libraryParams.asset
-    ? assets.filter((asset) => slugify(asset.content.name, { lower: true }) === libraryParams.asset)
+    ? assets.filter((asset) => getSlug(asset.content) === libraryParams.asset)
     : assets
 
   return {
     params: libraryParams,
     response,
     content: {
-      ...content,
+      ...packageJsonContent,
+      ...content, // spread last to use schema description if set
       private: !!content.private // default to false if not specified
     },
     assets: filteredAssets
@@ -104,18 +153,24 @@ export const getLibraryData = async (params = {}) => {
  * If the params map to a valid library in the allowlist, get the default branch if there isn't a
  * specified ref, then recursively get all asset metadata files. Find the files that are in the
  * library's subdirectory and then fetch the contents for each asset metadata file.
+ * @param {import('../typedefs').Params} params
+ * @param {boolean} inheritContent
+ * @returns {Promise<import('../typedefs').Asset[]>}
  */
-export const getLibraryAssets = async (params = {}) => {
+const getLibraryAssets = async (params = {}, inheritContent = false) => {
   const libraryParams = await validateLibraryParams(params)
 
-  if (!libraryParams || Object.keys(libraryParams).length === 0) return []
+  if (isEmpty(libraryParams)) return []
 
   // get all asset metadata files in subdirectories
 
-  let response = {}
+  /**
+   * @type {import('../typedefs').GitHubTreeResponse}
+   */
+  let treeResponse = {}
 
   try {
-    response = await getResponse(
+    treeResponse = await getResponse(
       libraryParams.host,
       'GET /repos/{owner}/{repo}/git/trees/{ref}?recursive=1',
       {
@@ -130,7 +185,7 @@ export const getLibraryAssets = async (params = {}) => {
 
   // request contents for each asset metadata file
 
-  const assetContentPromises = response.tree
+  const assetContentPromises = treeResponse.tree
     .filter(
       (file) =>
         removeLeadingSlash(file.path).startsWith(removeLeadingSlash(libraryParams.path)) &&
@@ -147,19 +202,144 @@ export const getLibraryAssets = async (params = {}) => {
 
   const assetContentData = await Promise.all(assetContentPromises)
 
-  const assets = assetContentData.map((response) => {
-    const content = yaml.load(Buffer.from(response.content, response.encoding).toString())
+  const assets = assetContentData
+    .map((response) => {
+      /**
+       * @type {import('../typedefs').AssetContent}
+       */
+      const content = yaml.load(Buffer.from(response.content, response.encoding).toString())
+
+      return {
+        params: libraryParams,
+        response,
+        content: {
+          ...content,
+          private: !!content.private // default to false if not specified
+        }
+      }
+    })
+    .filter((asset) => {
+      // if fetching a specific asset, only return that
+
+      return libraryParams.asset ? getSlug(asset.content) === libraryParams.asset : true
+    })
+
+  const inheritedAssets = inheritContent ? await getInheritedAssets(assets) : []
+
+  // TODO this is commented out to prevent production build errors
+  const imgPlaceholders = [] // await getImagePlaceholders(assets)
+
+  return assets.map((asset) => {
+    const assetExtensions = getAssetExtensions(asset, inheritedAssets, imgPlaceholders)
 
     return {
-      params: libraryParams,
-      response,
+      ...asset,
       content: {
-        ...content,
-        private: !!content.private // default to false if not specified
+        ...asset.content,
+        ...assetExtensions
+      }
+    }
+  })
+}
+
+/**
+ * Merges inherited content and placeholder image data into an asset's content.
+ * @param {import('../typedefs').Asset} originalAsset - Original asset
+ * @param {import('../typedefs').Asset[]} inheritedAssets - Inherited assets
+ * @param {import('../typedefs').PlaceholderImage[]} imgPlaceholders - Placeholder images
+ * @returns {import('../typedefs').AssetContent}
+ */
+const getAssetExtensions = (originalAsset, inheritedAssets, imgPlaceholders) => {
+  const assetExtensions = {}
+  const basePath = originalAsset.response.path.replace('/carbon-asset.yml', '')
+
+  // add image placeholder data
+
+  const foundImage = imgPlaceholders.find((image) => {
+    return (
+      image.img.src.includes(basePath) &&
+      image.img.src.includes(originalAsset.content.thumbnailPath)
+    )
+  })
+
+  if (foundImage) {
+    const { img, base64 } = foundImage
+    assetExtensions.thumbnailData = { img, base64 }
+  }
+
+  // add inherited data
+
+  if (originalAsset.content.inherits) {
+    const inheritedParams = getParamsFromInheritedAsset(originalAsset.content.inherits.asset)
+
+    const inheritedAsset = inheritedAssets.find((asset) => {
+      const params = asset.params
+
+      // if the inherited ref only specifies "latest", remove the ref from the original asset so
+      // we're not trying to match "latest" with the default branch "main" or "master"
+
+      if (!inheritedParams.ref) {
+        delete params.ref
+      }
+
+      return isEqual(params, inheritedParams)
+    })
+
+    if (
+      inheritedAsset &&
+      originalAsset.content.inherits &&
+      originalAsset.content.inherits.properties &&
+      originalAsset.content.inherits.properties.length
+    ) {
+      // loop over properties, extend
+
+      originalAsset.content.inherits.properties.forEach((property) => {
+        assetExtensions[property] = inheritedAsset.content[property] || ''
+      })
+
+      if (
+        originalAsset.content.inherits.properties.includes('thumbnailPath') &&
+        inheritedAsset.content.thumbnailData
+      ) {
+        assetExtensions.thumbnailData = inheritedAsset.content.thumbnailData
+      }
+    }
+  }
+
+  return assetExtensions
+}
+
+/**
+ * Iterates over an array of assets and returns an array of assets that are to be inherited.
+ * @param {import('../typedefs').Asset[]} assets - Assets
+ * @returns {Promise<import('../typedefs').Asset[]>} Array of assets that will be inherited
+ */
+const getInheritedAssets = async (assets) => {
+  const inheritedLibraryAssetsPromises = []
+
+  assets.forEach((asset) => {
+    if (asset.content.inherits) {
+      const inheritedParams = getParamsFromInheritedAsset(asset.content.inherits.asset)
+
+      if (!isEmpty(inheritedParams)) {
+        inheritedLibraryAssetsPromises.push(getLibraryAssets(inheritedParams, false))
       }
     }
   })
 
+  const inheritedAssetsResponses = await Promise.all(inheritedLibraryAssetsPromises)
+
+  return inheritedAssetsResponses.flat()
+}
+
+/**
+ * Finds thumbnail images, gets content of each image, writes each image to disk, generates
+ * placeholder images, returns placeholder images
+ * @param {import('../typedefs').Asset[]} assets - Assets
+ * @returns {Promise<import('../typedefs').PlaceholderImage[]>} Array of placeholder objects
+ */
+// eslint-disable-next-line no-unused-vars
+const getImagePlaceholders = async (assets) => {
   // find all thumbnail images, get content of each image, filter out assets with no thumbnail
 
   const imgContentsPromises = assets
@@ -199,37 +379,13 @@ export const getLibraryAssets = async (params = {}) => {
     })
   })
 
-  const imgPlaceholders = await Promise.all(imgPlaceholderPromises)
-
-  // merge in placeholder images
-
-  return assets.map((asset) => {
-    const assetExtensions = {}
-    const basePath = asset.response.path.replace('/carbon-asset.yml', '')
-
-    const foundImage = imgPlaceholders.find(
-      (image) =>
-        image.img.src.includes(basePath) && image.img.src.includes(asset.content.thumbnailPath)
-    )
-
-    if (foundImage) {
-      const { img, base64 } = foundImage
-      assetExtensions.thumbnailData = { img, base64 }
-    }
-
-    return {
-      ...asset,
-      content: {
-        ...asset.content,
-        ...assetExtensions
-      }
-    }
-  })
+  return Promise.all(imgPlaceholderPromises)
 }
 
 /**
  * Iterates over all libraries in the allowlist and fetches library data with no ref so the default
  * branch is used.
+ * @returns {import('../typedefs').Libraries}
  */
 export const getAllLibraries = async () => {
   const promises = []
@@ -248,5 +404,45 @@ export const getAllLibraries = async () => {
 
   return {
     libraries: libraries.filter((n) => n)
+  }
+}
+
+/**
+ * Requests content of the package.json file and returns some of the properties.
+ * @param {import('../typedefs').Params} params
+ * @param {string} packageJsonPath
+ * @returns {Promise<import('../typedefs').LibraryContent>}
+ */
+const getPackageJsonContent = async (params = {}, packageJsonPath = '/package.json') => {
+  const libraryParams = await validateLibraryParams(params)
+
+  if (isEmpty(libraryParams)) return {}
+
+  /**
+   * @type {import('../typedefs').GitHubContentResponse}
+   */
+  let response = {}
+
+  try {
+    response = await getResponse(libraryParams.host, 'GET /repos/{owner}/{repo}/contents/{path}', {
+      owner: libraryParams.org,
+      repo: libraryParams.repo,
+      path: removeLeadingSlash(`${libraryParams.path}${packageJsonPath}`),
+      ref: libraryParams.ref
+    })
+  } catch (err) {
+    return {}
+  }
+
+  /**
+   * @type {import('../typedefs').LibraryContent}
+   */
+  const packageJsonContent = yaml.load(Buffer.from(response.content, response.encoding).toString())
+
+  return {
+    description: packageJsonContent.description,
+    license: packageJsonContent.license,
+    package: packageJsonContent.name,
+    version: packageJsonContent.version
   }
 }